# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для платформы квестов Gengine",
        "title": "Gengine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gengine.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/games/{id}/export": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Выгружает уровни, вопросы и ответы игры в CSV-формат для редактирования или резервного копирования",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт игры в CSV",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV-файл с данными игры",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/games/{id}/export-excel": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Генерирует Excel-файл (.xlsx) со всеми уровнями, вопросами и ответами игры",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт игры в Excel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel-файл с данными игры",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/games/{id}/export-pdf": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Генерирует PDF-файл со всеми уровнями, вопросами и ответами игры для печати",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт игры в PDF",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF-файл с данными игры",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/games/{id}/export-results": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Выгружает итоговую таблицу результатов игры (места, команды, время, попытки) в CSV-формат",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт результатов в CSV",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV-файл с результатами",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/games/{id}/export-results-excel": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Генерирует Excel-файл (.xlsx) с итоговой таблицей результатов игры",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт результатов в Excel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel-файл с результатами",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/games/{id}/export-statistics-pdf": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Генерирует PDF-отчёт с расширенной статистикой игры (результаты команд по уровням: время, количество попыток)",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт статистики в PDF",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF-файл со статистикой",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/games/{id}/teams/{team_id}/export-results": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Выгружает результаты конкретной команды в CSV-формат (доступно капитану или автору игры)",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Экспорт результатов команды в CSV",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV-файл с результатами команды",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "jwt",
            "in": "cookie"
        }
    }
}